tidy up tower systems
