+++ core/src/towers/attacks/mod.rs
@@ -1,8 +1,6 @@
 mod basic_attack;
 mod speed_buff;
-mod towers_in_range;
 mod utils;
 
 pub use basic_attack::*;
 pub use speed_buff::*;
-pub use towers_in_range::*;
